Fixed Window focus logic
We now have the correct focus status of the app, meaning we can use the
controller in alerts and still deactivate controllers when in
background.
Does not work yet for custom Dialogs apparently, like GameRoomDialog.
But should not be an issue as we don't really need to use a controller
in it